\name{STstuff}
\alias{STstuff}
\title{The STstuff package}
\description{
	Package with the ST functions used for the metaclustering.
}
\details{
	Various functions for ST. Really.
}
\seealso{\link{STstuff}, \link{clustersInOtherDS}, \link{expressionInAnnotations},
  \link{deconvoluteClusters},  \link{computeET},  \link{computeMC}, \link{classifySpots}
}